requirements.txt
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
idna==2.8
imageio==2.5.0
joblib==0.13.2
kiwisolver==1.1.0
lxml==4.3.4
matplotlib==3.1.1
networkx==2.3
numpy==1.16.4
pandas==0.25.0
Pillow==6.1.0
pkg-resources==0.0.0
psycopg2==2.8.3
pyparsing==2.4.1.1
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
requests==2.22.0
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.2
SQLAlchemy==1.3.6
urllib3==1.25.3
